room users use player id

+++ Sadie.API/Game/Players/IPlayerRepository.cs
@@ -5,15 +5,15 @@ namespace Sadie.API.Game.Players;
 
 public interface IPlayerRepository
 {
-    IPlayerLogic? GetPlayerLogicById(int id);
+    IPlayerLogic? GetPlayerLogicById(long id);
     IPlayerLogic? GetPlayerLogicByUsername(string username);
-    Task<Player?> GetPlayerByIdAsync(int id);
+    Task<Player?> GetPlayerByIdAsync(long id);
     Task<Player?> GetPlayerByUsernameAsync(string username);
     ICollection<IPlayerLogic> GetAll();
     bool TryAddPlayer(IPlayerLogic player);
-    Task<bool> TryRemovePlayerAsync(int playerId);
-    int Count();
-    Task<List<Player>> GetPlayersForSearchAsync(string searchQuery, int[] excludeIds);
-    Task<List<PlayerRelationship>> GetRelationshipsForPlayerAsync(int playerId);
+    Task<bool> TryRemovePlayerAsync(long playerId);
+    long Count();
+    Task<List<Player>> GetPlayersForSearchAsync(string searchQuery, long[] excludeIds);
+    Task<List<PlayerRelationship>> GetRelationshipsForPlayerAsync(long playerId);
     Task BroadcastDataAsync(AbstractPacketWriter writer);
 }
