+++ Sadie.Database/Models/Group.cs
@@ -5,7 +5,7 @@ namespace Sadie.Database.Models;
 public class Group
 {
     public int Id { get; init; }
-    public int PlayerId { get; init; }
+    public long PlayerId { get; init; }
     public required string Name { get; init; }
     public required string Description { get; init; }
     public int RoomId { get; init; }
