+++ Sadie.Database/Models/Players/PlayerRoomLike.cs
@@ -3,6 +3,6 @@ namespace Sadie.Database.Models.Players;
 public class PlayerRoomLike
 {
     public int Id { get; init; }
-    public int PlayerId { get; init; }
+    public long PlayerId { get; init; }
     public int RoomId { get; init; }
 }
