+++ Sadie.Database/Models/Rooms/Chat/RoomChatMessage.cs
@@ -8,7 +8,7 @@ public class RoomChatMessage
 {
     public int Id { get; init; }
     public int RoomId { get; init; }
-    public int PlayerId { get; init; }
+    public long PlayerId { get; init; }
     public string? Message { get; init; }
     public ChatBubble ChatBubbleId { get; init; }
     public RoomUserEmotion EmotionId { get; init; }
