+++ Sadie.Game.Rooms/Packets/Writers/Users/HandItems/RoomUserHandItemWriter.cs
@@ -7,6 +7,6 @@ namespace Sadie.Game.Rooms.Packets.Writers.Users.HandItems;
 [PacketId(ServerPacketId.RoomUserHandItem)]
 public class RoomUserHandItemWriter : AbstractPacketWriter
 {
-    public required int UserId { get; init; }
+    public required long UserId { get; init; }
     public required int ItemId { get; init; }
 }
