+++ Sadie.Game.Rooms/Packets/Writers/Users/RoomUserWhisperWriter.cs
@@ -7,7 +7,7 @@ namespace Sadie.Game.Rooms.Packets.Writers.Users;
 [PacketId(ServerPacketId.RoomUserWhisper)]
 public class RoomUserWhisperWriter : AbstractPacketWriter
 {
-    public required int SenderId { get; set; }
+    public required long SenderId { get; set; }
     public required string Message { get; init; }
     public required int EmotionId { get; set; }
     public required int ChatBubbleId { get; set; }
