+++ Sadie.Networking.Writers/Rooms/Users/RoomUserDanceWriter.cs
@@ -6,6 +6,6 @@ namespace Sadie.Networking.Writers.Rooms.Users;
 [PacketId(ServerPacketId.RoomUserDance)]
 public class RoomUserDanceWriter : AbstractPacketWriter
 {
-    public required int UserId { get; init; }
+    public required long UserId { get; init; }
     public required int DanceId { get; init; }
 }
