+++ Sadie.Networking.Writers/Rooms/Users/RoomUserIdleWriter.cs
@@ -6,6 +6,6 @@ namespace Sadie.Networking.Writers.Rooms.Users;
 [PacketId(ServerPacketId.RoomUserIdle)]
 public class RoomUserIdleWriter : AbstractPacketWriter
 {
-    public required int UserId { get; set; }
+    public required long UserId { get; set; }
     public required bool IsIdle { get; set; }
 }
