+++ Sadie.Networking.Writers/Rooms/Users/Trading/RoomUserTradeStartedWriter.cs
@@ -7,14 +7,14 @@ namespace Sadie.Networking.Writers.Rooms.Users.Trading;
 [PacketId(ServerPacketId.RoomUserTradeStarted)]
 public class RoomUserTradeStartedWriter : AbstractPacketWriter
 {
-    public required List<int> UserIds { get; init; }
+    public required List<long> UserIds { get; init; }
     public required int State { get; init; }
 
     public override void OnSerialize(INetworkPacketWriter writer)
     {
         foreach (var id in UserIds)
         {
-            writer.WriteInteger(id);
+            writer.WriteLong(id);
             writer.WriteInteger(State);
         }
     }
